pyproject.toml
[project]
name = "a2a-multiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.6",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.6",
    "python-dotenv>=1.2.1",
]
